Removed old long-name project
For some reason when I renamed the project, it made a copy of it and renamed it. The old version was broken -- it just had include and src, none of the other necessary stuff to make it workable. Just tossed it because the new renamed copy is up to date and complete.